requirement.txt
fitz==0.0.1.dev2
numpy==1.23.5
pdf2image==1.17.0
pdfplumber==0.11.4
Pillow==11.0.0
pypdf==5.1.0
torch==2.4.1
tqdm==4.66.5
transformers==4.36.2
sentence-transformers==3.2.1
anthropic==0.39.0